implementing the CBT service
